// src/lib/api/generated/model/certificateDto.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * API
 * OpenAPI spec version: 1.0.0
 */
import type { CertificateStatus } from './certificateStatus';

export interface CertificateDto {
  id: string;
  serial: string;
  verifyCode: string;
  status: CertificateStatus;
  courseId: string;
  userId: string;
  /** @nullable */
  courseTitle?: string | null;
  issuedAt?: string;
}