// src/lib/api/generated/model/courseDetailDto.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * API
 * OpenAPI spec version: 1.0.0
 */
import type { CourseFaqDto } from './courseFaqDto';
import type { CourseLocalizationDto } from './courseLocalizationDto';
import type { CourseStatus } from './courseStatus';
import type { LanguageCode } from './languageCode';
import type { SectionDto } from './sectionDto';

export interface CourseDetailDto {
  id: string;
  instructorUserId: string;
  slug: string;
  status: CourseStatus;
  defaultLanguage: LanguageCode;
  /** @nullable */
  levelId?: string | null;
  disciplineIds: string[];
  specializationIds: string[];
  genreIds: string[];
  skillIds: string[];
  /** @nullable */
  thumbnailMediaId?: string | null;
  /** @nullable */
  promoVideoMediaId?: string | null;
  requirements: string[];
  audiences: string[];
  outcomes: string[];
  /** @nullable */
  priceAmount?: number | null;
  /** @nullable */
  priceCurrency?: string | null;
  /** @nullable */
  offeringId?: string | null;
  /** @nullable */
  publishedAt?: string | null;
  localizations: CourseLocalizationDto;
  sections: SectionDto;
  faqs: CourseFaqDto;
}