// src/lib/api/generated/model/lessonDto.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * API
 * OpenAPI spec version: 1.0.0
 */
import type { LessonAttachmentDto } from './lessonAttachmentDto';
import type { LessonCaptionDto } from './lessonCaptionDto';
import type { LessonType } from './lessonType';

export interface LessonDto {
  id: string;
  title: string;
  position: number;
  type: LessonType;
  /** @nullable */
  videoMediaId?: string | null;
  /** @nullable */
  textContent?: string | null;
  /** @nullable */
  durationSeconds?: number | null;
  isPreview: boolean;
  attachments: LessonAttachmentDto;
  captions: LessonCaptionDto;
}