// src/lib/api/generated/model/lessonProgressDto.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * API
 * OpenAPI spec version: 1.0.0
 */
import type { LessonProgressStatus } from './lessonProgressStatus';

export interface LessonProgressDto {
  lessonId: string;
  status: LessonProgressStatus;
  /** @nullable */
  lastPositionSeconds?: number | null;
  /** @nullable */
  completedAt?: string | null;
}